Allow controller subclasses to pick up template mappings defined in superclasses.
I noticed that if I define a template mapping with the template method, subclasses of the defining controller do not respect that mapping. This only really applies to the template :foo, :bar usage, where no explicit controller is specified. With these modifications, template mappings such as that will assume you mean the 'current' controller. Specs have been added to test this new functionality.