Improve action method resolving with respect to Helper::LOOKUP.
Now the action method resolving only considers helper methods where the helper is in Helper::LOOKUP and is included by the controller in question. Previously all methods of all helpers in Helper::LOOKUP were deemed to be available on all controllers whether they included the helpers or not, which was wrong.